Add call_option methods file Add display_list_option file Divide the method display_list and options from the main method
